# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/places": {
      "get": {
        "operationId": "CatalogController_searchPlaces",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Search points of interest (places) by text query or proximity",
        "tags": [
          "Catalog & Places"
        ]
      },
      "post": {
        "operationId": "CatalogController_suggestPlace",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Suggest a missing place (pending admin verification)",
        "tags": [
          "Catalog & Places"
        ]
      }
    },
...
    "/v1/driver/subscription": {
      "get": {
        "operationId": "SubscriptionsController_getDriverSubscription",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current driver active subscription and trial details",
        "tags": [
          "Driver Subscriptions"
        ]
      }
    },
...
    "/v1/admin/subscriptions": {
      "post": {
        "operationId": "SubscriptionsController_adminGrantSubscription",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Manually grant, activate, or extend driver subscription plan",
        "tags": [
          "Driver Subscriptions"
        ]
      }
    },
    "/v1/admin/subscriptions/{id}/payments": {
      "post": {
        "operationId": "SubscriptionsController_adminRecordPayment",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Record manual cash / offline payment for driver subscription",
        "tags": [
          "Driver Subscriptions"
        ]
      }
    },
...
    "/v1/agreements/{id}/messages": {
      "get": {
        "operationId": "MessagingController_getMessages",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retrieve conversation messages for an agreement (participants only)",
        "tags": [
          "In-App Messaging"
        ]
      },
      "post": {
        "operationId": "MessagingController_sendMessage",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send text or voice/image message within trip channel",
        "tags": [
          "In-App Messaging"
        ]
      }
    },
    "/v1/messages/{id}/read": {
      "post": {
        "operationId": "MessagingController_markRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark message as read",
        "tags": [
          "In-App Messaging"
        ]
      }
    },
...
    "/v1/agreements/{id}/ratings": {
      "post": {
        "operationId": "RatingsController_createRating",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit mutual evaluation (1-5 stars) after trip completion",
        "tags": [
          "Trust, Ratings & Disputes"
        ]
      }
    },
...
    "/v1/disputes": {
      "post": {
        "operationId": "RatingsController_createDispute",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "File a dispute or claim regarding an order or agreement",
        "tags": [
          "Trust, Ratings & Disputes"
        ]
      },
      "get": {
        "operationId": "RatingsController_listDisputes",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List filed disputes (users see own cases; admins see all)",
        "tags": [
          "Trust, Ratings & Disputes"
        ]
      }
    },
    "/v1/admin/disputes/{id}/assign": {
      "post": {
        "operationId": "RatingsController_assignDispute",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign support agent or administrator to investigate dispute",
        "tags": [
          "Trust, Ratings & Disputes"
        ]
      }
    },
    "/v1/admin/disputes/{id}/events": {
      "post": {
        "operationId": "RatingsController_addDisputeEvent",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add note or event to dispute investigation log",
        "tags": [
          "Trust, Ratings & Disputes"
        ]
      }
    },
    "/v1/admin/disputes/{id}/resolve": {
      "post": {
        "operationId": "RatingsController_resolveDispute",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Resolve or dismiss dispute with audit recording",
        "tags": [
          "Trust, Ratings & Disputes"
        ]
      }
    },
...
  },
  "info": {
    "title": "Wasel API",
    "description": "Wasel Logistics & Delivery Platform Modular Monolith API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "idempotency-key": {
        "type": "apiKey",
        "in": "header",
        "name": "Idempotency-Key"
      }
    },
...
  }
}